% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addm2_fixation_model_random.R
\name{addm2_fixation_model_random}
\alias{addm2_fixation_model_random}
\title{Fixation model example (for usage with aevacc functions)}
\usage{
addm2_fixation_model_random()
}
\arguments{
\item{fixdur.vec}{Vector of fixation-durations to sample from}
}
\value{
Returns matrix with row one providing fixation location and row two providing fixation duration
}
\description{
Compute seqeuence of fixation locations and durations (random locations)
\code{addm_fixation_model_random} Returns matrix with row one providing fixation location and row two providing fixation duration
#'
}
\author{
Alexander Fengler, \email{alexanderfengler@gmx.de}
}

