% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addm2_plot_family.R
\name{addm2_plot_family}
\alias{addm2_plot_family}
\title{Convenience Function for generating standard plots}
\usage{
addm2_plot_family(choice.dat = data.table(v1 = 0, v2 = 0, rt = 0, decision =
  0, id = 0), eye.dat = data.table(fixloc = 0, fixnr = 1, fixdur = 0, id = 0),
  addm.output = data.table(a = 0, b = 0))
}
\arguments{
\item{choice.dat}{A 'data.frame' or  'data.table' storing the item valuations (v1,v2...), reaction times in ms (rt), decisions as decision and an id column (id). A by trial form is assumed.}

\item{eye.dat}{A 'data.frame' or 'data.table' storing eyetracking data by trial. Fixation location (fixloc), Fixation number (fixnr), Fixation duration (fixdur) and an id column (id). Can all be initialized as zero columns when a by condition fit is attempted.}

\item{addm.output}{data.table that stores detailed addm.output (as generated by addm_run_by_condition)}
}
\value{
Returns a list of ggplot objects
}
\description{
Provides a few standard plots that are generally useful when analyzing addm fits
\code{addm_plot_family} Returns list with a few diagnostic plots
}
\author{
Alexander Fengler, \email{alexanderfengler@gmx.de}
}

