% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addm_boundary_concave.R
\name{addm_boundary_concave}
\alias{addm_boundary_concave}
\title{Concave boundary}
\usage{
addm_boundary_concave(maxrt = 3000, timestep = 10,
  time_decay_start = 1000, decay = 1.001)
}
\arguments{
\item{maxrt}{integer providing the maximum rt that should be considered for the boundary}

\item{timestep}{integer that provides the timestep-size that the current fit utilized}

\item{time_decay_start}{time delay for start of decay (boundary parameter)}

\item{decay}{numeric variable providing value for decay concavity}
}
\value{
vector that provides boundary levels by timestep
\code{addm_boundary_concave()}
}
\description{
Generate boundaries with concave decay, and potential decay delay
}
\author{
Alexander Fengler, \email{alexanderfengler@gmx.de}
}

