% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addm_by_condition.R
\name{addm_by_condition}
\alias{addm_by_condition}
\title{Model run by condition}
\usage{
addm_by_condition(conditions.dat, eye.dat, choice.dat, model.parameters,
  nr.reps, model.type, output.type, fixation.model, timestep, generate)
}
\arguments{
\item{conditions.dat}{A 'data.frame' storing the item valuations (v1,v2...) by unique trial conditions. An id column (id) needs to be provided that matches by trial data.}

\item{eye.dat}{A 'data.frame' or 'data.table' storing eyetracking data by trial. Fixation location (fixloc), Fixation number (fixnr), Fixation duration (fixdur) and an id column (id). Can all be initialized as zero columns when a by condition fit is attempted.}

\item{choice.dat}{A 'data.frame' or  'data.table' storing the item valuations (v1,v2...) , reaction times in ms (rt), decisions as decision and an id column (id). A by trial form is assumed.}

\item{model.parameters}{A vector with the four core addm parameters in order (drift.rate,theta,sd,non.decision.time).}

\item{nr.reps}{An integer number that tells the function how many simulation runs to use.}

\item{model.type}{A string that indicates which version of the model to run. 'standard' or 'memnoise' when memory effects shall be allowed.}

\item{output.type}{A string that indicates what output the model shall produce. 'full' for detailed model output, 'fit' for sparse output (rt,decision) by id variable.}

\item{fixation.model}{A string that indicates which fixation model will be utilized for simulations. 'random' for random fixations (implemented) 'fakepath' for following a predetermined fixation path with fixed durations (implemented)}

\item{timestep}{An integer number that provides the timestep-size that is used in the simulations (in ms).}

\item{generate}{Binary variable that tells the function to return either log likelihood values (0) or rt, decision (1). Relevant only if model.type variable is 'fit'.}
}
\value{
The function has three potential return values. A log likelihood value, utilized when trying to fit the model. A simple data.table providing simulated rts,decisions by condition id, which useful for generating fake data in testing. A full model output with many details, utilized for running the model with optimal parameters and extracting data for plots.
\code{addm_by_condition} Returns either log likelihoods, a simple id,decision,rt data frame or a detailed model output
}
\description{
Runs the model by unique trial conditions.
}
\author{
Alexander Fengler, \email{alexanderfengler@gmx.de}
}

