% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addm_dataprep.R
\name{addm_dataprep}
\alias{addm_dataprep}
\title{Data Preparation for easy supply to fit functions}
\usage{
addm_dataprep(choice.dat = data.table(v1 = 0, v2 = 0, id = 0, rt = 0, decision
  = 0, id = 0), eye.dat = data.table(fixloc = 0, fixnr = 1, fixdur = 0, id =
  0), timestep = 10, rtbinsize = 100)
}
\arguments{
\item{choice.dat}{A 'data.frame' or  'data.table' storing the item valuations (v1,v2...) , reaction times in ms (rt), decisions as decision and an id column (id). A by trial form is assumed.}

\item{eye.dat}{A 'data.frame' or 'data.table' storing eyetracking data by trial. Fixation location (fixloc), Fixation number (fixnr), Fixation duration (fixdur) and an id column (id). Can all be initialized as zero columns when a by condition fit is attempted.}

\item{timestep}{An integer number that provides the timestep-size that is used in the simulations (in ms).}

\item{rtbinsize}{An integer number that provides the binsize that is used for binning reaction time data (in ms).}

\item{fit.type}{String defined as 'condition' or 'trial' for the respective model fits}
}
\value{
returns a list of three components. A 'data.table' that stores all unique choice set conditions. A 'data.table' that stores eyetracking.data adjusted to be usable for by trial fits. A 'data.table' that stores by trial choice data. An id variable links all three data.tables.
}
\description{
Shapes up a 'data.frame' or 'data.table' to be straight usable with the other addm_*** functions
\code{addm_dataprep} List of three components. A 'data.table' that stores all unique choice set conditions. A 'data.table' that stores eyetracking.data adjusted to be usable for by trial fits. A 'data.table' that stores by trial choice data. An id variable links all three data.tables.
}
\author{
Alexander Fengler, \email{alexanderfengler@gmx.de}
}

