% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addm_fit_grid.R
\name{addm_fit_grid}
\alias{addm_fit_grid}
\title{Run Gridsearch over supplied parameter space}
\usage{
addm_fit_grid(data = list(choice.dat = NULL, eye.dat = NULL, conditions.dat =
  NULL), drifts = seq(5e-04, 0.003, 5e-04), thetas = seq(0, 1, 0.25),
  sds = seq(0.05, 0.09, 0.01), non.decision.times = 0, nr.reps = 2000,
  timestep = 10, model.type = "standard", output.type = "fit",
  fixation.model = "fixedpath", fit.type = "condition",
  allow.fine.grid = 0, coarse.to.fine.ratio = 4,
  log.file = "defaultlog.txt", parallel = 1)
}
\arguments{
\item{data}{A list of three data.tables of each: choice data, eyetracking data, conditions data (as created by addm_dataprep)}

\item{drifts}{Vector of all driftrate values to be tested.}

\item{thetas}{Vector of all theta to be tested (between 0 and 1).}

\item{sds}{Vector of all standard deviation values to be tested.}

\item{non.decision.times}{Vector of all non decision times to be tested (in ms).}

\item{nr.reps}{An integer number that tells the function how many simulation runs to use.}

\item{timestep}{An integer number that provides the timestep-size that is used in the simulations (in ms).}

\item{model.type}{A string that indicates which version of the model to run. 'standard' or 'memnoise' when memory effects shall be allowed.}

\item{output.type}{A string that indicates what output the model shall produce. 'full' for detailed model output, 'fit' for sparse output (rt,decision) by id variable.}

\item{fixation.model}{A string that indicates which fixation model will be utilized for simulations. 'random' for random fixations (implemented) 'fixedpath' for following a predetermined fixation path with fixed durations (implemented).}

\item{fit.type}{A string indicating either 'condition' for fits by unique trial conditions or 'trial' for fits by trial.}

\item{allow.fine.grid}{Binary variable that indicates whether we allow (1) a fine grid to be created an searched around the coarse grid minimum or not (0).}

\item{coarse.to.fine.ratio}{Integer number deifining the ratio between parameter steps in the coarse vs. the fine grid. Defaults to 4.}

\item{log.file}{Filepath for storage of logs}

\item{parallel}{Binary variable that tells whether to initialize local cluster on start (1) or not (0).}
}
\value{
Returns data.table with log likelihoods by parameter combination
\code{addm_fit_grid} Returns data.table with log likelihoods by parameter combination
}
\description{
Runs grid search over supplied parameter space
}
\author{
Alexander Fengler, \email{alexanderfengler@gmx.de}
}

