% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addm_fixation_model_example.R
\name{addm_fixation_model_example}
\alias{addm_fixation_model_example}
\title{Fixation model example (for usage with aevacc functions)}
\usage{
addm_fixation_model_example(last.fixloc = 0, fixdur.vec = 0)
}
\arguments{
\item{last.fixloc}{Numeric Variable providing last fixation location in simulation}

\item{fixdur.vec}{Vector of fixation-durations to sample from}
}
\value{
Returns vector with [1] fixation location and [0] fixation duration
}
\description{
Compute new fixation location and duration
\code{addm_gridsearch_foreach} Returns vector with [1] fixation location and [0] fixation duration
#'
}
\author{
Alexander Fengler, \email{alexanderfengler@gmx.de}
}

