% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addm_fixation_model_fixedpath.R
\name{addm_fixation_model_fixedpath}
\alias{addm_fixation_model_fixedpath}
\title{Example for user-supplied fixation model (aevacc functions)}
\usage{
addm_fixation_model_fixedpath(cur.valuations = c(0, 0), set.size = 2)
}
\arguments{
\item{cur.valuations}{vector of valuations in current choice set}

\item{set.size}{number of items present in current choice set}
}
\value{
matrix with two rows. One providing fixation location and the second providing fixation duration
}
\description{
Compute seqeuence of fixation locations and durations (fixed locations)
\code{addm_fixation_model_fixedpath}
}
\author{
Alexander Fengler, \email{alexanderfengler@gmx.de}
}

