% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addm_generate_artificial.R
\name{addm_generate_artificial}
\alias{addm_generate_artificial}
\title{Generate artificial dataset}
\usage{
addm_generate_artificial(set.size = 2, possible.valuations = c(0, 1, 2, 3),
  model.parameters = c(0.002, 0.5, 0.07, 0), nr.attributes = 1,
  nr.reps = 100, timestep = 10, rtbinsize = 100, nr.conditions = 10,
  model.type = "standard")
}
\arguments{
\item{set.size}{Indicates the number of items that are allowed.}

\item{possible.valuations}{Vector storing valuations that single items can hold in a choice set.}

\item{model.parameters}{vector of parameters that will be used by the model to generate artificial choice/rt data}

\item{nr.attributes}{integer providing the number of attributes that will be assigned to each item in the artificial choice set}

\item{nr.reps}{An integer number that tells the function how many simulation runs to use.}

\item{timestep}{An integer number that provides the timestep-size that is used in the simulations (in ms).}

\item{rtbinsize}{integer providing the binsize that reaction times will be sorted into}

\item{nr.conditions}{integer that provides the number of unique trial conditions to be generated}

\item{model.type}{A string that indicates which version of the model to run. 'standard' or 'memnoise' when memory effects shall be allowed.}

\item{fixation.model}{A string that indicates which fixation model will be utilized for simulations. 'random' for random fixations (implemented) 'fixedpath' for following a predetermined fixation path with fixed durations (implemented).}

\item{core.parameters}{Vector that provide parameters used to generate artificial data from drift diffusion process. (drift,theta,sd,non deicision time)}
}
\value{
returns a list of three components. A 'data.table' that stores all unique choice set conditions. A 'data.table' that stores eyetracking.data adjusted to be usable for by trial fits. A 'data.table' that stores by trial choice data. An id variable links all three data.tables.
\code{addm_generate_artificial} Returns artificial data that can be used for test addm model fits
}
\description{
Generate artificial data for addm model fits
}
\author{
Alexander Fengler, \email{alexanderfengler@gmx.de}
}

