% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addm_generate_by_condition_artificial.R
\name{addm_generate_by_condition_artificial}
\alias{addm_generate_by_condition_artificial}
\title{Generate artificial dataset}
\usage{
addm_generate_by_condition_artificial(set.size = 2,
  possible.valuations = c(0, 1, 2, 3), timestep = 10, nr.reps = 2000,
  model.type = "standard", fixation.model = "fixedpath",
  core.parameters = c(0.006, 0.6, 0.06, 0))
}
\arguments{
\item{set.size}{Indicates the number of items that are allowed.}

\item{possible.valuations}{Vector storing valuations that single items can hold in a choice set.}

\item{timestep}{An integer number that provides the timestep-size that is used in the simulations (in ms).}

\item{nr.reps}{An integer number that tells the function how many simulation runs to use.}

\item{model.type}{A string that indicates which version of the model to run. 'standard' or 'memnoise' when memory effects shall be allowed.}

\item{fixation.model}{A string that indicates which fixation model will be utilized for simulations. 'random' for random fixations (implemented) 'fixedpath' for following a predetermined fixation path with fixed durations (implemented).}

\item{core.parameters}{Vector that provide parameters used to generate artificial data from drift diffusion process. (drift,theta,sd,non deicision time)}
}
\value{
returns a list of three components. A 'data.table' that stores all unique choice set conditions. A 'data.table' that stores eyetracking.data adjusted to be usable for by trial fits. A 'data.table' that stores by trial choice data. An id variable links all three data.tables.
\code{addm_generate_by_condition_artificial} Returns artificial data that can be used for test addm model fits
}
\description{
Generate artificial data for addm model fits
}
\author{
Alexander Fengler, \email{alexanderfengler@gmx.de}
}

