% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addm_plot_loglik.R
\name{addm_plot_loglik}
\alias{addm_plot_loglik}
\title{Plot modelfit}
\usage{
addm_plot_loglik(logliks)
}
\arguments{
\item{logliks}{'data.frame' or 'data.table' storing parameter values and corresponding log likelihoods as a result of a model fit.}
}
\value{
Returns a list of two ggplot objects. plot.coarse.grid stores a plot corresponding to coarse grid fits. plot.fine.grid analogously for fine grid fits
\code{addm_plot_loglik} Returns data.table with log likelihoods by parameter combination
}
\description{
Produce model fit visualization
}
\author{
Alexander Fengler, \email{alexanderfengler@gmx.de}
}

