% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addm_run_by_trial_dynamic.R
\name{addm_run_by_trial_dynamic}
\alias{addm_run_by_trial_dynamic}
\title{Run model by unique trial}
\usage{
addm_run_by_trial_dynamic(choice.dat = data.table(v1 = 0, v2 = 0, id = 0),
  eye.dat = data.table(fixloc = 0, fixdur = 0, fixnr = 1, id = 0),
  model.parameters = c(0.002, 0.5, 0.07, 0), timestep = 10,
  state.step = 0.1)
}
\arguments{
\item{choice.dat}{data.table storing the item valuations (v1,v2...), reaction times in ms (rt), decisions (decision) and unique trial ids (id).}

\item{eye.dat}{data.table storing eyetracking data by trial. Fixation location (fixloc), Fixation number (fixnr), Fixation duration (fixdur) and a unique trial ids (id).}

\item{model.parameters}{vector with the four core addm parameters in order (drift.rate, theta, sd, non.decision.time).}

\item{timestep}{integer number that provides the timestep-size that is used in the simulations (in ms).}

\item{state.step}{numeric variable between 0 and 1, that indicate how finegrained the vertial grid is supposed to be}
}
\value{
numeric variable storing the negative log likelihood for given parameter set
\code{addm_run_by_trial_dynamic}
}
\description{
Run model by unique trial for one set of parameter values
}
\author{
Alexander Fengler, \email{alexanderfengler@gmx.de}
}

