% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addm_support_gridsearch_foreach.R
\name{addm_support_gridsearch_foreach}
\alias{addm_support_gridsearch_foreach}
\title{Run grid search (parallel)}
\usage{
addm_support_gridsearch_foreach(choice.dat = data.table(v1 = 0, v2 = 0, rt =
  0, decision = 0, id = 0), eye.dat = data.table(fixloc = 0, fixdur = 0, fixnr
  = 1, id = 0), conditions.dat = data.table(v1 = 0, v2 = 0, id = 0),
  parameter.matrix = c(0.006, 0.6, 1, 0.06, 0), nr.attributes = 1,
  nr.reps = 1000, timestep = 10, model.type = "standard",
  fixation.model = "fixedpath", fit.type = "condition",
  log.file = "defaultlog.txt", state.step = 0.1)
}
\arguments{
\item{parameter.matrix}{matrix that provides the parameter space which is looped over. (drifts, thetas, sds, non decision times / potentially gammas when using multiattribute case)}

\item{nr.attributes}{integer providing the amount of attributes we consider per item}

\item{nr.reps}{integer that tells the function how many simulation runs to use.}

\item{timestep}{integer that provides the timestep-size that is used in the simulations (in ms).}

\item{model.type}{string that indicates which version of the model to run. 'standard' for normal model fits. 'memnoise' to allow for memory effects (see vignette for more for detailed explanation of what this is about).}

\item{fixation.model}{string that indicates which fixation model will be utilized for simulations. 'random' for random fixations (example). 'fixedpath' for following a predetermined fixation path with fixed durations (example). 'user' to provide your own fixation model, defined in a function "user_fixation_model" in the global environment.}

\item{fit.type}{string indicating either 'condition' for fits by unique trial conditions, 'trial' for fits by trial, or 'dyn' where you can use a dynamic programming algorithm for fitting the two items case, bypassing simulations for the fitting procedure}

\item{log.file}{path to a file for storing fit-logs}

\item{state.step}{parameter only relevant when using fit.type = 'dyn', for which case it given the precision of the vertical grid utilized in the dynammic programming algorithm}
}
\value{
data.table with log likelihoods and corresponding parameter combinations
}
\description{
Run (parallel) grid search for addm fits using foreach loops
\code{addm_gridsearch_foreach}
}
\author{
Alexander Fengler, \email{alexanderfengler@gmx.de}
}

