% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{aevacc2_by_condition}
\alias{aevacc2_by_condition}
\title{Simulate aDDM process (by condition, 2 items)
\code{aevacc2_by_condition()}}
\usage{
aevacc2_by_condition(parameters, maxdur, update, nr_attributes, fixation_model,
  nr_reps, timestep)
}
\arguments{
\item{parameters}{vector that stores the parameters used for the simulations (Order: [non.decision.time, drift, sd, theta])}

\item{maxdur}{maximum duration in ms that the process is allowed to simulate}

\item{update}{Vector that stores the item valuations for the trial conditon simulated}

\item{nr_attributes}{placeholder for interface consistency / see multiattribute versions for specification}

\item{fixation_model}{a user supplied fixation model that will be utilized to supply fixation locations and potentially fixation durations}

\item{nr_reps}{number of repitions (simulation runs)}

\item{timestep}{timestep in ms associated with each step in the drift diffusion process}
}
\value{
vector that stores decisions and rts for each simulation run
}
\description{
Simulate aDDM process by unique trial condition (2 items)
}
\author{
Alexander Fengler, \email{alexanderfengler@gmx.de}
}

