% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{aevacc_by_condition}
\alias{aevacc_by_condition}
\title{Evidence accumulation by condition (item general)
\code{aevacc_by_condition}}
\usage{
aevacc_by_condition(sd, theta, drift, non_decision_time, timestep, nr_reps,
  maxdur, update, fixpos, fixdur, fixdursamples, fixation_model)
}
\arguments{
\item{sd}{standard deviation used for drift diffusion process}

\item{theta}{theta used for drift diffusion process}

\item{drift}{drift-rate used for drift diffusion process}

\item{non_decision_time}{non decision time used for drift diffusion process}

\item{timestep}{timestep in ms associated with each step in the drift diffusion process}

\item{nr_reps}{number of repitions (simulation runs)}

\item{maxdur}{maximum duration in ms that the process is allowed to simulate}

\item{update}{Vector that stores the item valuations for the trial conditon simulated}

\item{fixpos}{Vector that stores the locations for a supplied fixed fixation pathway}

\item{fixdur}{Vector that stores the fixation durations for a supplied fixed fixation pathway}

\item{fixdursamples}{Vector from which fixation duration can be sampled once supplied fixations run out}

\item{fixation_model}{a user supplied fixation model that will be utilized to supply fixation locations and potentially fixation durations}
}
\value{
Returns a vector that stores decisions and rts for each simulation run
}
\description{
Runs evidence accumulation function (item general case) for one trial condition
}
\author{
Alexander Fengler, \email{alexanderfengler@gmx.de}
}

