% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{aevacc_full_output_memnoise}
\alias{aevacc_full_output_memnoise}
\title{Simulate aDDM process (>2 items, allow memory effects) for detailed model output
\code{aevacc_full_output_memnoise()}}
\usage{
aevacc_full_output_memnoise(parameters, maxdur, update, fixation_model, nr_reps,
  timestep)
}
\arguments{
\item{parameters}{vector that stores the parameters used for the simulations (Order: [non.decision.time, drift, sd, theta, gamma, scalar_items_seen_drift, scala_items_seen_noise])}

\item{maxdur}{maximum duration in ms that the process is allowed to simulate}

\item{update}{Vector that stores the item valuations for the trial conditon simulated}

\item{fixation_model}{a user supplied fixation model that will be utilized to supply fixation locations and potentially fixation durations}

\item{nr_reps}{number of repitions (simulation runs)}

\item{timestep}{timestep in ms associated with each step in the drift diffusion process}
}
\value{
vector that stores detailed output by simulation run
}
\description{
Simulate aDDM process (>2 items, allow memory effects) for detailed model output
}
\author{
Alexander Fengler, \email{alexanderfengler@gmx.de}
}

