% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{aevaccma2_by_trial}
\alias{aevaccma2_by_trial}
\title{Simulate aDDM process (by trial, 2 items, multiattribute)}
\usage{
aevaccma2_by_trial(sd, theta, gamma, drift, non_decision_time, maxdur, mindur,
  cur_decision, update, nr_attributes, fixpos, fixdur, nr_reps, timestep)
}
\arguments{
\item{sd}{standard deviation used for drift diffusion process}

\item{theta}{theta (attentional bias) used for drift diffusion process [0,1]}

\item{gamma}{secondary attentional bias which refers to attributes no inspected [0,1]}

\item{drift}{drift-rate used for drift diffusion process}

\item{non_decision_time}{non decision time used for drift diffusion process}

\item{maxdur}{numeric variable that supplies the maximum reaction time considered a success in simulations}

\item{mindur}{numeric variable that supplies the minimum reaction time considered a succes in simulations}

\item{cur_decision}{numeric variable that provides the empirical decision taken in trial}

\item{update}{Vector that stores the item valuations for the trial conditon simulated}

\item{fixpos}{Vector that stores the locations for a supplied fixed fixation pathway}

\item{fixdur}{Vector that stores the fixation durations for a supplied fixed fixation pathway}

\item{nr_reps}{number of repitions (simulation runs)}

\item{timestep}{timestep in ms associated with each step in the drift diffusion process}

\item{cur_maxfix}{integer that provides number of fixation in trial}
}
\value{
numeric variable that provides a success count (runs that predicted a reaction time in the correct rt-bin and simultaneously the correct decision)
}
\description{
Simulate aDDM process by unique trial (2 items // multiattribute)
\code{aevaccma2_by_trial()}
}
\author{
Alexander Fengler, \email{alexanderfengler@gmx.de}
}

