% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dynamicaddm}
\alias{dynamicaddm}
\title{Simulate aDDM process (by condition, 2 items)
\code{dynamicaddm()}}
\usage{
dynamicaddm(sd, theta, drift, non_decision_time, decision, valuations, fixpos,
  fixdur, rt, stateStep)
}
\arguments{
\item{sd}{standard deviation used for drift diffusion process}

\item{theta}{theta used for drift diffusion process}

\item{drift}{drift-rate used for drift diffusion process}

\item{non_decision_time}{non decision time used for drift diffusion process}

\item{decision}{integer which gives choice in current trial (1 or 2)}

\item{valuations}{vector that stores the item valuations for the provided trial}

\item{fixpos}{vector with all empirical fixations positions encountered in provided trial}

\item{fixdur}{vector with all empirical fixation durations encounrtered in provided trial}

\item{rt}{reaction time of provided trial}

\item{stateStep}{numeric variable between [0,1] that indicates how finegrained the vertical grid of the model space shall be computed}
}
\value{
numeric variable storing likelihood value
}
\description{
Simulate aDDM process by unique trial condition (2 items)
}
\author{
Alexander Fengler, \email{alexanderfengler@gmx.de}
}

