% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dynamicddm}
\alias{dynamicddm}
\title{Simulate DDM process (by condition, 2 items)
\code{dynamicddm()}}
\usage{
dynamicddm(parameters, decision = 0L, valuations = 0L, nr_attributes = 0L,
  fixpos = 0L, fixdur = 0L, rt = 0L, stateStep = 0)
}
\arguments{
\item{decision}{integer which gives choice in current trial (1 or 2)}

\item{valuations}{vector that stores the item valuations for the provided trial}

\item{nr_attributes}{integer variable placeholder for interface consistency / see multiattribute versions for specification}

\item{fixpos}{vector placeholder for interface consistency / see attentional versions for specification}

\item{fixdur}{vector placeholder for interface consistency / see attentional versions for specification}

\item{rt}{reaction time of provided trial}

\item{stateStep}{numeric variable between [0,1] that indicates how finegrained the vertical grid of the model space shall be computed}
}
\value{
numeric variable storing likelihood value
}
\description{
Simulate DDM process by unique trial condition (2 items)
}
\author{
Alexander Fengler, \email{alexanderfengler@gmx.de}
}

