% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{evacc2_by_trial}
\alias{evacc2_by_trial}
\title{Simulate DDM process (by trial, 2 items)}
\usage{
evacc2_by_trial(sd = 0, theta = 0, gamma = 0, drift = 0,
  non_decision_time = 0L, maxdur = 0L, mindur = 0L, cur_decision = 0L,
  update = 0L, nr_attributes = 0L, fixpos = 0L, fixdur = 0L,
  nr_reps = 0L, timestep = 0L)
}
\arguments{
\item{sd}{standard deviation used for drift diffusion process}

\item{theta}{placeholder for interface consistency / see attentional versions for specification}

\item{gamma}{placeholder for interface consistency / see multiattribute versions for specification}

\item{drift}{drift-rate used for drift diffusion process}

\item{non_decision_time}{non decision time used for drift diffusion process}

\item{maxdur}{numeric variable that supplies the maximum reaction time considered a success in simulations}

\item{mindur}{numeric variable that supplies the minimum reaction time considered a succes in simulations}

\item{cur_decision}{numeric variable that provides the empirical decision taken in trial}

\item{update}{Vector that stores the item valuations for the trial conditon simulated}

\item{nr_attributes}{placeholder for interface consistency / see multiattribute versions for specification}

\item{fixpos}{Vector placeholder for interface consistency / see attentional versions for specification}

\item{fixdur}{Vector placeholder for interface consistency / see attentional versions for specification}

\item{nr_reps}{number of repitions (simulation runs)}

\item{timestep}{timestep in ms associated with each step in the drift diffusion process}

\item{cur_maxfix}{integer that provides number of fixation in trial}
}
\value{
numeric variable that provides a success count (runs that predicted a reaction time in the correct rt-bin and simultaneously the correct decision)
}
\description{
Simulate DDM process by unique trial (2 items)
\code{evacc2_by_trial()}
}
\author{
Alexander Fengler, \email{alexanderfengler@gmx.de}
}

