% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addm2_fixation_model_fixedpath.R
\name{addm2_fixation_model_fixedpath}
\alias{addm2_fixation_model_fixedpath}
\title{Example for user-supplied fixation model (aevacc functions)}
\usage{
addm2_fixation_model_fixedpath(cur.valuations = c(0, 0), fixdur = 250)
}
\arguments{
\item{cur.valuations}{vector of current valuations in choice set}

\item{fixdur}{integer variable that provides the duration of fixations simulated (in ms)}
}
\value{
matrix with two rows. One providing fixation location and the second providing fixation duration
}
\description{
Compute seqeuence of fixation locations and durations (fixed locations)
\code{addm2_fixation_model_fixedpath}
}
\author{
Alexander Fengler, \email{alexanderfengler@gmx.de}
}

