% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addm2_fixation_model_random.R
\name{addm2_fixation_model_random}
\alias{addm2_fixation_model_random}
\title{Example for user-supplied fixation model (aevacc functions)}
\usage{
addm2_fixation_model_random(cur.valuations = c(0, 0))
}
\arguments{
\item{cur.valuations}{vector that provides item valuations in current choice set}
}
\value{
matrix with two rows. One providing fixation location and the second providing fixation duration
}
\description{
Compute seqeuence of fixation locations and durations (random locations)
\code{addm2_fixation_model_random}
}
\author{
Alexander Fengler, \email{alexanderfengler@gmx.de}
}

