% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addm2_plot_family.R
\name{addm2_plot_family}
\alias{addm2_plot_family}
\title{Generate standard psychometric plots for addm model-fit analyis}
\usage{
addm2_plot_family(choice.dat = data.table(v1 = 0, v2 = 0, rt = 0, decision =
  0, id = 0), eye.dat = data.table(fixloc = 0, fixnr = 1, fixdur = 0, id = 0),
  addm.output = data.table(a = 0, b = 0))
}
\arguments{
\item{choice.dat}{data.table storing the item valuations (v1,v2...), reaction times in ms (rt), decisions (decision), trialids (id) and condition ids (condition_id).}

\item{eye.dat}{data.table storing eyetracking data by trial. Fixation location (fixloc), Fixation number (fixnr), Fixation duration (fixdur) and an id column (id). Not strictly neccessary at the moment (not plot currently uses it) and can be run as default.}

\item{addm.output}{data.table that stores detailed addm.output (as generated by addm_run_by_condition)}
}
\value{
List of ggplot objects (plots). Costumizable.
}
\description{
Generate standard psychometric plots for addm model-fit analysis
\code{addm_plot_family}
}
\author{
Alexander Fengler, \email{alexanderfengler@gmx.de}
}

