% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addm_dataprep.R
\name{addm_dataprep}
\alias{addm_dataprep}
\title{Data Preprocessing for easy usage of addmtoolbox package}
\usage{
addm_dataprep(choice.dat = data.table(v1 = 0, v2 = 0, id = 0, rt = 0, decision
  = 0, id = 0), eye.dat = data.table(fixloc = 0, fixnr = 1, fixdur = 0, id =
  0), timestep = 10, rtbinsize = 100)
}
\arguments{
\item{choice.dat}{data.frame or data.table storing the item valuations (v1,v2...) , reaction times in ms (rt), decisions (decision) and a trial id column (id).}

\item{eye.dat}{A data.frame or data.table storing eyetracking data by trial. Fixation location (fixloc), Fixation number (fixnr), Fixation duration (fixdur) and a trial id column (id). Can all be initialized as zero columns when a by condition fit is attempted.}

\item{timestep}{integer that provides the timestep-size that is used in the simulations (in ms).}

\item{rtbinsize}{integer that provides the binsize that is used for binning reaction time data (in ms).}

\item{fit.type}{string defined as 'condition' or 'trial' for the respective model fits}
}
\value{
list of three components. data.table that stores all unique choice set conditions. data.table that stores eyetracking.data adjusted to be usable for by trial fits. data.table that stores by trial choice data. trial id and condition id variable link all data.tables
}
\description{
Preprocess experimental data for easy usage with addmtoolbox package
\code{addm_dataprep}
}
\author{
Alexander Fengler, \email{alexanderfengler@gmx.de}
}

