% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addm_fit_grid.R
\name{addm_fit_grid}
\alias{addm_fit_grid}
\title{Run Gridsearch over supplied parameter space}
\usage{
addm_fit_grid(conditions.dat = data.table(v1 = c(1, 2, 3), v2 = c(3, 2, 1), id
  = c(1, 2, 3)), eye.dat = data.table(fixloc = 0, fixnr = 0, fixdur = 0, id =
  c(1, 2, 3)), choice.dat = data.table(v1 = c(1, 2, 3), v2 = c(3, 2, 1), id =
  c(1, 2, 3), rt = c(0, 0, 0), decision = c(1, 1, 1)), drifts = seq(0.002,
  0.01, 0.002), thetas = seq(0.2, 1, 0.2), sds = seq(0.02, 0.1, 0.02),
  non.decision.times = 0, timestep = 10, nr.reps = 2000,
  model.type = "standard", output.type = "fit", fit.type = "condition",
  fixation.model = "fakepath", allow.fine.grid = 0,
  log.file = "defaultlog.txt")
}
\arguments{
\item{conditions.dat}{A 'data.frame' storing the item valuations (v1,v2...) by unique trial conditions. An id column (id) needs to be provided that matches by trial data.}

\item{eye.dat}{A 'data.frame' or 'data.table' storing eyetracking data by trial. Fixation location (fixloc), Fixation number (fixnr), Fixation duration (fixdur) and an id column (id). Can all be initialized as zero columns when a by condition fit is attempted.}

\item{choice.dat}{A 'data.frame' or  'data.table' storing the item valuations (v1,v2...) , reaction times in ms (rt), decisions as decision and an id column (id). A by trial form is assumed.}

\item{drifts}{Vector of all driftrate values to be tested.}

\item{thetas}{Vector of all theta to be tested (between 0 and 1).}

\item{sds}{Vector of all standard deviation values to be tested.}

\item{non.decision.times}{Vector of all non decision times to be tested (in ms).}

\item{timestep}{An integer number that provides the timestep-size that is used in the simulations (in ms).}

\item{nr.reps}{An integer number that tells the function how many simulation runs to use.}

\item{model.type}{A string that indicates which version of the model to run. 'standard' or 'memnoise' when memory effects shall be allowed.}

\item{output.type}{A string that indicates what output the model shall produce. 'full' for detailed model output, 'fit' for sparse output (rt,decision) by id variable.}

\item{fit.type}{A string indicating either 'condition' for fits by unique trial conditions or 'trial' for fits by trial.}

\item{fixation.model}{A string that indicates which fixation model will be utilized for simulations. 'random' for random fixations (implemented) 'fakepath' for following a predetermined fixation path with fixed durations (implemented).}

\item{allow.fine.grid}{Binary variable that indicates whether we allow (1) a fine grid to be created an searched around the coarse grid minimum or not (0).}

\item{log.file}{Filepath for storage of logs}
}
\value{
Returns data.table with log likelihoods by parameter combination
\code{addm_fit_grid} Returns data.table with log likelihoods by parameter combination
}
\description{
Runs grid search over supplied parameter space
}
\author{
Alexander Fengler, \email{alexanderfengler@gmx.de}
}

