% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addm_fit_grid.R
\name{addm_fit_grid}
\alias{addm_fit_grid}
\title{Fit addm using grid search}
\usage{
addm_fit_grid(data = list(choice.dat = NULL, eye.dat = NULL, conditions.dat =
  NULL), drifts = seq(0.001, 0.0025, 5e-04), thetas = seq(0, 1, 0.25),
  sds = seq(0.05, 0.09, 0.01), non.decision.times = 0, nr.reps = 1000,
  timestep = 10, model.type = "standard", fixation.model = "fixedpath",
  fit.type = "condition", allow.fine.grid = 0, coarse.to.fine.ratio = 4,
  log.file = "defaultlog.txt", parallel = 1, state.step = 0.1)
}
\arguments{
\item{data}{list of three data.tables of each: choice data, eyetracking data, conditions data (as created by addm_dataprep)}

\item{drifts}{vector of all driftrate values to be tested.}

\item{thetas}{vector of all theta to be tested (between 0 and 1).}

\item{sds}{vector of all standard deviation values to be tested.}

\item{non.decision.times}{vector of all non decision times to be tested (in ms).}

\item{nr.reps}{integer that tells the function how many simulation runs to use.}

\item{timestep}{integer that provides the timestep-size that is used in the simulations (in ms).}

\item{model.type}{string that indicates which version of the model to run. 'standard' for normal model fits. 'memnoise' to allow for memory effects (see vignette for more for detailed explanation of what this is about).}

\item{fixation.model}{string that indicates which fixation model will be utilized for simulations. 'random' for random fixations (example). 'fixedpath' for following a predetermined fixation path with fixed durations (example). 'user' to provide your own fixation model, defined in a function "user_fixation_model" in the global environment.}

\item{fit.type}{string indicating either 'condition' for fits by unique trial conditions, 'trial' for fits by trial, or 'dyn' where you can use a dynamic programming algorithm for fitting the two items case, bypassing simulations for the fitting procedure}

\item{allow.fine.grid}{variable that indicates whether we allow (1) a fine grid to be created and searched around the coarse grid minimum or not (0).}

\item{coarse.to.fine.ratio}{integer defining the ratio between parameter steps in the coarse versus the fine grid.}

\item{log.file}{path to a file for storing fit-logs}

\item{parallel}{boolean varible that indicates whether to initialize local cluster on start (1) or not (0).}

\item{state.step}{parameter only relevant when using fit.type = 'dyn', for which case it given the precision of the vertical grid utilized in the dynammic programming algorithm}
}
\value{
data.table with log likelihoods by parameter combination
\code{addm_fit_grid}
}
\description{
Run grid search over supplied parameter space
}
\author{
Alexander Fengler, \email{alexanderfengler@gmx.de}
}

