% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addm_plot_loglik.R
\name{addm_plot_loglik}
\alias{addm_plot_loglik}
\title{Visualize Likelihood fitting}
\usage{
addm_plot_loglik(logliks = data.table(drift = 0, theta = 0, sd = 0,
  non.decision.time = 0, nr.reps = 0, loglik = 0))
}
\arguments{
\item{logliks}{'data.frame' or 'data.table' storing parameter values and corresponding log likelihoods as a result of a model fit.}
}
\value{
list of two ggplot objects. plot.coarse.grid stores a plot corresponding to coarse grid fits. plot.fine.grid analogously stores a plot for fine grid fits.
\code{addm_plot_loglik}
}
\description{
Visualize Likelihood fitting
}
\author{
Alexander Fengler, \email{alexanderfengler@gmx.de}
}

