% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addm_support_compute_finegrid.R
\name{addm_support_compute_finegrid}
\alias{addm_support_compute_finegrid}
\title{Generate fine grid around best parameter combination}
\usage{
addm_support_compute_finegrid(drift.step.fine, sd.step.fine, thetas,
  non.decision.time.step.fine, coarse.to.fine.ratio, log.liks)
}
\arguments{
\item{drift.step.fine}{Numeric variable storing precision with which drift rate space is searched in the fine grid parameter matrix.}

\item{sd.step.fine}{Numeric variable storing precision with which standard deviation space is searched in the fine grid parameter matrix.}

\item{thetas}{Numeric variable storing testable thetas. The theta space is not changed in the fine parameter matrix.}

\item{non.decision.time.step.fine}{Numeric variable storing precision with which non.decision.time space is searched in the fine grid parameter matrix.}

\item{coarse.to.fine.ratio}{Numeric Variable that stores the precision ratio between the coarse and fine grid for grid search fits.}

\item{log.liks}{'data.table' that stores log likelihoods and corresponding parameter combinations resulting from coarse grid search.}
}
\value{
Matrix storing parameter space
}
\description{
Function internal to grid search, which computes a fine grid around the coarse grid minimum log likelihood
\code{addm_support_compute_finegrid} Gives back a parameter matrix
#'
}
\author{
Alexander Fengler, \email{alexanderfengler@gmx.de}
}

