% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{aevacc2_by_condition}
\alias{aevacc2_by_condition}
\title{evidence accumulation (2 item case)
\code{aevacc2_by_condition} Returns data.table with log likelihoods and corresponding parameter combinations}
\usage{
aevacc2_by_condition(sd, theta, drift, non_decision_time, maxdur, update,
  fixation_model, nr_reps, timestep)
}
\arguments{
\item{sd}{standard deviation used for drift diffusion process}

\item{theta}{theta used for drift diffusion process}

\item{drift}{drift-rate used for drift diffusion process}

\item{non_decision_time}{non decision time used for drift diffusion process}

\item{maxdur}{maximum duration in ms that the process is allowed to simulate}

\item{update}{Vector that stores the item valuations for the trial conditon simulated}

\item{fixation_model}{a user supplied fixation model that will be utilized to supply fixation locations and potentially fixation durations}

\item{nr_reps}{number of repitions (simulation runs)}

\item{timestep}{timestep in ms associated with each step in the drift diffusion process}
}
\value{
Returns a vector that stores decisions and rts for each simulation run
}
\description{
Runs evidence accumulation function (2 item case) for one trial condition
}
\author{
Alexander Fengler, \email{alexanderfengler@gmx.de}
}

