% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{aevacc2_by_trial}
\alias{aevacc2_by_trial}
\title{Simulate aDDM process (by trial, 2 items)}
\usage{
aevacc2_by_trial(parameters, maxdur = 0L, mindur = 0L, cur_decision = 0L,
  update = 0L, nr_attributes = 0L, fixpos = 0L, fixdur = 0L,
  nr_reps = 0L, timestep = 0L)
}
\arguments{
\item{parameters}{vector that stores the parameters used for the simulations (Order: [non.decision.time, drift, sd, theta])}

\item{maxdur}{numeric variable that supplies the maximum reaction time considered a success in simulations}

\item{mindur}{numeric variable that supplies the minimum reaction time considered a succes in simulations}

\item{cur_decision}{numeric variable that provides the empirical decision taken in trial}

\item{update}{Vector that stores the item valuations for the trial conditon simulated}

\item{nr_attributes}{placeholder for interface consistency / see multiattribute versions for specification}

\item{fixpos}{Vector that stores the locations for a supplied fixed fixation pathway}

\item{fixdur}{Vector that stores the fixation durations for a supplied fixed fixation pathway}

\item{nr_reps}{number of repitions (simulation runs)}

\item{timestep}{timestep in ms associated with each step in the drift diffusion process}

\item{cur_maxfix}{integer that provides number of fixation in trial}
}
\value{
numeric variable that provides a success count (runs that predicted a reaction time in the correct rt-bin and simultaneously the correct decision)
}
\description{
Simulate aDDM process by unique trial (2 items)
\code{aevacc2_by_trial()}
}
\author{
Alexander Fengler, \email{alexanderfengler@gmx.de}
}

