% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{aevacc2_by_trial}
\alias{aevacc2_by_trial}
\title{evidence accumulation (2 item case) by trial}
\usage{
aevacc2_by_trial(sd, theta, drift, non_decision_time, maxdur, mindur,
  real_decision, update, fixpos, fixdur, nr_reps, timestep)
}
\arguments{
\item{sd}{standard deviation used for drift diffusion process}

\item{theta}{theta used for drift diffusion process}

\item{drift}{drift-rate used for drift diffusion process}

\item{non_decision_time}{non decision time used for drift diffusion process}

\item{maxdur}{numeric variable that supplies the maximum reaction time considered a success in simulations}

\item{mindur}{numeric variable that supplies the minimum reaction time considered a succes in simulations}

\item{real_decision}{numeric variable that provides the empirical decision taken in trial}

\item{update}{Vector that stores the item valuations for the trial conditon simulated}

\item{fixpos}{Vector that stores the locations for a supplied fixed fixation pathway}

\item{fixdur}{Vector that stores the fixation durations for a supplied fixed fixation pathway}

\item{nr_reps}{number of repitions (simulation runs)}

\item{timestep}{timestep in ms associated with each step in the drift diffusion process}
}
\value{
Returns a numeric variable that provides a success count (runs that predicted a reaction time in the correct rt-bin and simultaneously the correct decision)
}
\description{
Runs evidence accumulation function (2 item case) for one trial condition
\code{aevacc2_by_trial}
}
\author{
Alexander Fengler, \email{alexanderfengler@gmx.de}
}

