% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{aevacc_by_condition}
\alias{aevacc_by_condition}
\title{Simulate aDDM process by unique trial condition (>2 items)
\code{aevacc_by_condition()}}
\usage{
aevacc_by_condition(sd, theta, drift, non_decision_time, maxdur, update,
  fixation_model, nr_reps, timestep)
}
\arguments{
\item{sd}{standard deviation used for drift diffusion process}

\item{theta}{theta (attentional bias) used for drift diffusion process}

\item{drift}{drift-rate used for drift diffusion process}

\item{non_decision_time}{non decision time used for drift diffusion process}

\item{maxdur}{maximum duration in ms that the process is allowed to simulate}

\item{update}{Vector that stores the item valuations for the trial conditon simulated}

\item{fixation_model}{a user supplied fixation model that will be utilized to supply fixation locations and potentially fixation durations}

\item{nr_reps}{number of repitions (simulation runs)}

\item{timestep}{timestep in ms associated with each step in the drift diffusion process}

\item{fixpos}{Vector that stores the locations for a supplied fixed fixation pathway}

\item{fixdur}{Vector that stores the fixation durations for a supplied fixed fixation pathway}

\item{fixdursamples}{Vector from which fixation duration can be sampled once supplied fixations run out}
}
\value{
vector that stores decisions and rts for each simulation run
}
\description{
Simulate aDDM process by unique trial condition (>2 items)
}
\author{
Alexander Fengler, \email{alexanderfengler@gmx.de}
}

