% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{aevacc_full_output_memnoise}
\alias{aevacc_full_output_memnoise}
\title{Evidence accumulation by condition (item general) with detailed output
\code{aevacc_full_output_memnoise}}
\usage{
aevacc_full_output_memnoise(sd, theta, drift, non_decision_time, timestep,
  nr_reps, maxdur, update, fixation_model)
}
\arguments{
\item{sd}{standard deviation used for drift diffusion process}

\item{theta}{theta used for drift diffusion process}

\item{drift}{drift-rate used for drift diffusion process}

\item{non_decision_time}{non decision time used for drift diffusion process}

\item{timestep}{timestep in ms associated with each step in the drift diffusion process}

\item{nr_reps}{number of repitions (simulation runs)}

\item{maxdur}{maximum duration in ms that the process is allowed to simulate}

\item{update}{Vector that stores the item valuations for the trial conditon simulated}

\item{fixation_model}{a user supplied fixation model that will be utilized to supply fixation locations and potentially fixation durations}
}
\value{
Returns a vector that stores detailed output by simulation run
}
\description{
Runs evidence accumulation function (item general case) for one trial, and returns detailed model output
}
\author{
Alexander Fengler, \email{alexanderfengler@gmx.de}
}

