% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{aevaccma2_by_condition}
\alias{aevaccma2_by_condition}
\title{Simulate aDDM process (by condition, 2 items)
\code{aevaccma2_by_condition()}}
\usage{
aevaccma2_by_condition(sd = 0, theta = 0, gamma = 0, drift = 0,
  non_decision_time = 0L, maxdur = 0L, update = 0L, nr_attributes = 1L,
  fixation_model = 0L, nr_reps = 0L, timestep = 0L)
}
\arguments{
\item{sd}{standard deviation used for drift diffusion process}

\item{theta}{theta used for drift diffusion process [0,1]}

\item{gamma}{secondary attentional bias which refers to attributes no inspected [0,1]}

\item{drift}{drift-rate used for drift diffusion process}

\item{non_decision_time}{non decision time used for drift diffusion process}

\item{maxdur}{maximum duration in ms that the process is allowed to simulate}

\item{update}{Vector that stores the item valuations for the trial conditon simulated}

\item{fixation_model}{a user supplied fixation model that will be utilized to supply fixation locations and potentially fixation durations}

\item{nr_reps}{number of repitions (simulation runs)}

\item{timestep}{timestep in ms associated with each step in the drift diffusion process}
}
\value{
vector that stores decisions and rts for each simulation run
}
\description{
Simulate aDDM process by unique trial condition (2 items, multiattribute)
}
\author{
Alexander Fengler, \email{alexanderfengler@gmx.de}
}

