% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dynamicmaaddm}
\alias{dynamicmaaddm}
\title{Simulate aDDM process (by condition, 2 items)
\code{dynamicaddm()}}
\usage{
dynamicmaaddm(sd = 0, theta = 0, gamma = 0, drift = 0,
  non_decision_time = 0L, decision = 0L, valuations = 0L,
  nr_attributes = 0L, fixpos = 0L, fixdur = 0L, rt = 0L,
  stateStep = 0)
}
\arguments{
\item{sd}{standard deviation used for drift diffusion process}

\item{theta}{theta used for drift diffusion process [0,1]}

\item{gamma}{secondary attentional bias which refers to attributes no inspected [0,1]}

\item{drift}{drift-rate used for drift diffusion process}

\item{non_decision_time}{non decision time used for drift diffusion process}

\item{decision}{integer which gives choice in current trial (1 or 2)}

\item{valuations}{vector that stores the item valuations for the provided trial}

\item{nr_attributes}{integer which gives the number of attributes by item}

\item{fixpos}{vector with all empirical fixations positions encountered in provided trial}

\item{fixdur}{vector with all empirical fixation durations encounrtered in provided trial}

\item{rt}{reaction time of provided trial}

\item{stateStep}{numeric variable between [0,1] that indicates how finegrained the vertical grid of the model space shall be computed}
}
\value{
numeric variable storing likelihood value
}
\description{
Simulate aDDM process by unique trial condition (2 items)
}
\author{
Alexander Fengler, \email{alexanderfengler@gmx.de}
}

