% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{evacc2_by_condition}
\alias{evacc2_by_condition}
\title{Simulate DDM process (by condition, 2 items)
\code{evacc2_by_condition()}}
\usage{
evacc2_by_condition(sd, theta, gamma, drift, non_decision_time, maxdur, update,
  nr_attributes, fixation_model, nr_reps, timestep)
}
\arguments{
\item{sd}{standard deviation used for drift diffusion process}

\item{theta}{placeholder for interface consistency / see attentional versions for specification}

\item{gamma}{placeholder for interface consistency / see multiattribute versions for specification}

\item{drift}{drift-rate used for drift diffusion process}

\item{non_decision_time}{non decision time used for drift diffusion process}

\item{maxdur}{maximum duration in ms that the process is allowed to simulate}

\item{update}{Vector that stores the item valuations for the trial conditon simulated}

\item{nr_attributes}{placeholder for interface consistency / see multiattribute versions for specification}

\item{fixation_model}{placeholder for interface consistency / see attentional versions for specification}

\item{nr_reps}{number of repitions (simulation runs)}

\item{timestep}{timestep in ms associated with each step in the drift diffusion process}
}
\value{
vector that stores decisions and rts for each simulation run
}
\description{
Simulate DDM process by unique trial condition (2 items)
}
\author{
Alexander Fengler, \email{alexanderfengler@gmx.de}
}

